requirements.txt
amqp==2.1.4
appdirs==1.4.3
billiard==3.5.0.2
celery==4.0.2
Django==1.10.6
djangorestframework==3.6.2
kombu==4.0.2
MySQL-python==1.2.5
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
vine==1.1.3
